+++ R/checks.R
@@ -22,3 +22,13 @@ check_binary <- function(x, name_x = "x") {
 
   invisible(x)
 }
+
+
+check_not_null <- function(x, name_x = "x") {
+  if (is.null(x)) {
+    rlang::abort(paste0("`", name_x,
+                        "` must be specified and can't be `NULL`."))
+  }
+
+  invisible(x)
+}