+++ R/scimo-package.R
@@ -2,7 +2,7 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom generics tidy
+#' @importFrom generics required_pkgs tidy
 #' @importFrom recipes prep bake
 #' @importFrom tibble tibble
 ## usethis namespace: end